+++ src/main/parsing.ino
@@ -7,7 +7,7 @@
 
 String parsing(String data, char separator, int index)
 {
-  int found = 0;
+  uint8_t found = 0;
   int strIndex[] = {0, -1};
   int maxIndex = data.length()-1;
   for(int i=0; i<=maxIndex && found<=index; i++){
